+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 -----
 - Fixed product variation language bug
 - Fixed measurement language bug
-- Added product variation type support
+- Fixed category loading by name
+- Added product variation type (standard, swatches) support
 - Added product child image relation support
 
 1.1.1